+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/CallbackQueryReceivers/CallbackQueryReceiver.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.utils.CallbackQueryReceivers
+
+import com.pengrad.telegrambot.TelegramBot
+import com.pengrad.telegrambot.model.CallbackQuery
+import java.lang.ref.WeakReference
+
+interface CallbackQueryReceiver : (CallbackQuery, TelegramBot) -> Unit